+++ contracts/reflect/examples/schema.rs
@@ -2,7 +2,7 @@ use std::env::current_dir;
 use std::fs::create_dir_all;
 
 use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
-use cosmwasm_std::HandleResponse;
+use cosmwasm_std::Response;
 
 use reflect::msg::{
     CapitalizedResponse, ChainResponse, CustomMsg, HandleMsg, InitMsg, OwnerResponse, QueryMsg,
@@ -19,7 +19,7 @@ fn main() {
     export_schema(&schema_for!(CustomMsg), &out_dir);
     export_schema(&schema_for!(InitMsg), &out_dir);
     export_schema(&schema_for!(HandleMsg), &out_dir);
-    export_schema(&schema_for!(HandleResponse<CustomMsg>), &out_dir);
+    export_schema(&schema_for!(Response<CustomMsg>), &out_dir);
     export_schema(&schema_for!(QueryMsg), &out_dir);
     export_schema(&schema_for!(State), &out_dir);