+++ packages/std/src/coins.rs
@@ -23,13 +23,13 @@ impl Coin {
 /// # Examples
 ///
 /// ```
-/// # use cosmwasm_std::{coins, BankMsg, CosmosMsg, HandleResponse};
+/// # use cosmwasm_std::{coins, BankMsg, CosmosMsg, Response};
 /// # use cosmwasm_std::testing::{mock_env, mock_info};
 /// # let env = mock_env();
 /// # let info = mock_info("sender", &[]);
 /// let tip = coins(123, "ucosm");
 ///
-/// let mut response: HandleResponse = Default::default();
+/// let mut response: Response = Default::default();
 /// response.messages = vec![CosmosMsg::Bank(BankMsg::Send {
 ///   to_address: info.sender,
 ///   amount: tip,
@@ -44,7 +44,7 @@ pub fn coins<S: Into<String>>(amount: u128, denom: S) -> Vec<Coin> {
 /// # Examples
 ///
 /// ```
-/// # use cosmwasm_std::{coin, BankMsg, CosmosMsg, HandleResponse};
+/// # use cosmwasm_std::{coin, BankMsg, CosmosMsg, Response};
 /// # use cosmwasm_std::testing::{mock_env, mock_info};
 /// # let env = mock_env();
 /// # let info = mock_info("sender", &[]);
@@ -53,7 +53,7 @@ pub fn coins<S: Into<String>>(amount: u128, denom: S) -> Vec<Coin> {
 ///     coin(24, "ustake"),
 /// ];
 ///
-/// let mut response: HandleResponse = Default::default();
+/// let mut response: Response = Default::default();
 /// response.messages = vec![CosmosMsg::Bank(BankMsg::Send {
 ///     to_address: info.sender,
 ///     amount: tip,